requirements.txt
Pillow==7.2.0
mido==1.2.9
python-rtmidi==1.4.3
pygame==1.9.6
